requirements.txt
capstone==5.0.1
keystone==25.0.0
lief==0.15.1
pwntools==4.12.0
